text: split sharedparley into modules
One 2600-line file held five separable concerns whose boundaries were
enforced by comments. Splitting them apart makes the boundaries visible in
the module tree and every crossing explicit as a `pub(super)`:

- `shaping`: from text to shaped paragraphs. `shape_paragraphs` is the one
  function every path shapes through, so that measuring and drawing register
  identical cache dependencies.
- `cache`: the shaped-paragraph cache, its RAII guard, and the checkout of
  entries that only `with_text_layout` performs.
- `layout`: line breaking, alignment, the elision and max-lines cuts, and
  the queries the entry points ask of a `Layout`.
- `draw`: the `GlyphRenderer` trait and the painting of glyph runs,
  decorations and the clip-based recoloring of selected glyphs.
- `selection`: selection geometry, resolved into the per-line spans whose
  rounded edges both fill the highlight and clip the glyphs.

The root keeps the public facade -- unchanged for external users -- along
with `with_text_layout` and the entry points. Pure code motion plus module
docs and visibility annotations; no behavior change.
